app/requirements.txt
# HuggingFace Spaces Deployment Requirements
# For ZeroGPU deployment

# Core ML
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0

# Embeddings and Vector Store
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4

# Web Interface
gradio>=4.44.0

# ZeroGPU
spaces>=0.20.0

# Utilities
numpy>=1.24.3
tqdm>=4.66.1
